Refactored SettingForm component to split custom hooks and other external function into a separate utils.js
